PAP_2222123_WalissonGandorini_Finalizada-main/project/backend/requirements.txt
# Core FastAPI framework
fastapi>=0.111.0
uvicorn>=0.29.0

# Security and Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Supabase Client
supabase>=2.5.0

# Other utilities
python-multipart==0.0.6
python-dotenv>=1.0.0
requests>=2.31.0

# Music and Data Processing
music21==8.1.0
numpy>=1.26.0 # Upgraded for Python 3.12 compatibility
# No longer pinning pydantic, letting fastapi handle its version.

# PDF and Image Processing (kept versions for stability)
PyPDF2==3.0.1
Pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
pdfminer.six==20221105
opencv-python

# Azure SDK
azure-cognitiveservices-vision-customvision

# --- Testing ---
pytest
httpx # For making requests to the app in tests
